+++ README.md
@@ -16,3 +16,4 @@
 - [x] 가변인자를 사용해 PreparedStatementSetter 인터페이스 구현의 번거로움 낮추기
 - [x] 함수 분리 및 정리
 - [x] DatabasePopulatorUtils 리팩터링
+- [x] jdbc 패키지 jdbc 모듈로 이동
