+++ jdbc/src/main/java/nextstep/jdbc/exception/ExecuteQueryException.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao.jdbc.template.exception;
+package nextstep.jdbc.exception;
 
 public class ExecuteQueryException extends RuntimeException {
