+++ jdbc/src/main/java/nextstep/jdbc/exception/ExecuteUpdateException.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao.jdbc.template.exception;
+package nextstep.jdbc.exception;
 
 public class ExecuteUpdateException extends RuntimeException {
